added boolean to check for deletion

+++ dfirtrack_main/importer/file/filesystem.py
@@ -1,6 +1,7 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect
+from dfirtrack.config import REPORTITEMS_DELETE as reportitems_delete
 from dfirtrack.config import REPORTITEMS_FILESYSTEMPATH as reportitems_filesystempath
 from dfirtrack.config import REPORTITEMS_HEADLINE as reportitems_headline
 from dfirtrack.config import REPORTITEMS_SUBHEADLINE as reportitems_subheadline
@@ -47,6 +48,14 @@ def reportitems(request):
         # leave importer
         return redirect('/systems/')
 
+    # check whether REPORTITEMS_DELETE is defined in `dfirtrack.config`
+    if not isinstance(reportitems_delete, bool):
+        # call logger
+        error_logger(str(request.user), " REPORTITEMS_DELETE_VARIABLE_UNDEFINED")
+        messages.error(request, "The variable REPORTITEMS_DELETE seems to be undefined or not a boolean. Check `dfirtrack.config`!")
+        # leave importer
+        return redirect('/systems/')
+
     # get all system objects
     systems = System.objects.all()
 
@@ -80,16 +89,20 @@ def reportitems(request):
             warning_logger(str(request.user), " REPORTITEMS_FILESYSTEM_IMPORTER_NO_FILE system_name:" + system.system_name)
             # autoincrement counter
             nofile_found_counter += 1
-            # delete already existing reportitem for this system if no file was provided
-            try:
-                reportitem = Reportitem.objects.get(system = system, headline = headline, reportitem_subheadline = reportitems_subheadline)
-                # call logger (before deleting instance)
-                reportitem.logger(str(request.user), " REPORTITEMS_FILESYSTEM_IMPORTER_REPORTITEM_DELETED")
-                reportitem.delete()
-                # autoincrement counter
-                reportitems_deleted_counter += 1
-            except Reportitem.DoesNotExist:
-                pass
+
+            # check whether already existing reportitem for this system should be deleted if no file was provided
+            if reportitems_delete:
+                # delete already existing reportitem for this system if no file was provided
+                try:
+                    reportitem = Reportitem.objects.get(system = system, headline = headline, reportitem_subheadline = reportitems_subheadline)
+                    # call logger (before deleting instance)
+                    reportitem.logger(str(request.user), " REPORTITEMS_FILESYSTEM_IMPORTER_REPORTITEM_DELETED")
+                    reportitem.delete()
+                    # autoincrement counter
+                    reportitems_deleted_counter += 1
+                except Reportitem.DoesNotExist:
+                    pass
+
             # continue with next system
             continue
